% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpop-package.R
\docType{package}
\name{ggpopdata-package}
\alias{ggpopdata}
\alias{ggpopdata-package}
\title{ggpopdata: Data for Visualizing Population Data (ggpop)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ggpop is an R package that creates visually engaging and informative population charts. ggpop allows users to represent population data proportionally using customizable icons, facilitating the creation of circular representative population charts.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/jurjoroa/ggpopdata/issues}
}

}
\author{
\strong{Maintainer}: Jorge Roa \email{jorgeroa@stanford.edu} (\href{https://orcid.org/0000-0002-3972-9793}{ORCID})

}
\keyword{internal}
