% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load coordinatesDataset from .rda File}
\usage{
load_data()
}
\value{
A data frame containing the loaded dataset.

A data.frame containing df_coordinates_final
}
\description{
This function loads a dataset from an \code{.rda} file named
\code{df_coordinates_final.rda} and returns it.
}
\examples{
\dontrun{
# Load the dataset
df_coordinates <- load_data()

# Inspect the first few rows
head(df_coordinates)
}
}
